# src/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "yofun API Doc",
        "description": "Description about simple API methods in yofun project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1": {
            "get": {
                "tags": [
                    "main"
                ],
                "summary": "Get data for index page",
                "operationId": "61e9ead6a8de556b9adec5d67f20c342",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sliderEvents": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnnouncementResource"
                                            }
                                        },
                                        "upcomingEvents": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnnouncementResource"
                                            }
                                        },
                                        "upcomingQuizzes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnnouncementResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/quizzes": {
            "get": {
                "tags": [
                    "main"
                ],
                "summary": "Get all quizzes",
                "operationId": "458e7264bc57853a017e6989c08786f2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Get items in target page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnnouncementResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/events": {
            "get": {
                "tags": [
                    "main"
                ],
                "summary": "Get all events",
                "operationId": "7271709104caff040d6878140dcf20a4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Get items in target page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnnouncementResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/static": {
            "get": {
                "tags": [
                    "main"
                ],
                "summary": "Get static data",
                "operationId": "167addd4a04453229e87bdf3f28a3bde",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "todayCount": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "totalCount": {
                                            "type": "integer",
                                            "example": 23
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/feedback": {
            "post": {
                "tags": [
                    "main"
                ],
                "summary": "Send feedback",
                "operationId": "91a76d8a0cba1d93fedc5d9e08e6fc99",
                "requestBody": {
                    "description": "Feedback data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name (min 2 characters)",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Valid email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "subject": {
                                        "description": "Message subject",
                                        "type": "string",
                                        "example": "Frontend features"
                                    },
                                    "message": {
                                        "description": "Feedback text content",
                                        "type": "string",
                                        "example": "This is my feedback message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Name must have at least 4 letters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AnnouncementResource": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "Трибьютный концерт"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание трибьютного концерта"
                    },
                    "detail": {
                        "type": "string",
                        "example": "example.com"
                    },
                    "date": {
                        "type": "string",
                        "example": "2025-06-21"
                    },
                    "time": {
                        "type": "string",
                        "example": "20:00"
                    },
                    "location": {
                        "type": "string",
                        "example": "паб Бульдог"
                    },
                    "price": {
                        "type": "number",
                        "example": "1500"
                    },
                    "category": {
                        "type": "string",
                        "example": "концерты"
                    },
                    "image": {
                        "type": "string",
                        "example": "example.com/image.png"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "main",
            "description": "main"
        }
    ]
}